Traktor: iOS fixes, mostly OpenGL ES 2 renderer issues but also touch input and memory configuration.
[git-p4: depot-paths = "//depot/traktor/": change = 6589]

+++ code/Core/Memory/FastAllocator.cpp
@@ -67,8 +67,8 @@ void* FastAllocator::alloc(size_t size, size_t align, const char* const tag)
 
 		size = nearestLog2(size);
 
-		size_t qid = log2(size);
-		size_t qsize = 1 << qid;
+		uint32_t qid = log2(size);
+		uint32_t qsize = 1UL << qid;
 
 		qid -= 4;
